refactor(webviews)!: rename _integration folder for sort-order clarity
Renames src/webviews/webviewIntegration/ to src/webviews/_integration/.
The underscore prefix is the conventional "infrastructure / not feature
code" signal in monorepos, and it sorts before feature folders
(components/, documentdb/, ...) in VS Code's file explorer. With api/
the folder was first by alphabetical accident; with webviewIntegration/
it sorted at the bottom. _integration restores the desired top placement
while staying honest about the role (local glue between this extension
and @microsoft/vscode-ext-react-webview).

Breaking change for in-tree consumers; the npm package surface is
unchanged. All imports across src/ and the skill docs under .github/
are updated.
